requirements.txt
Flask==0.12.1
google-cloud-pubsub==0.25.0
gunicorn==19.7.1
webapp2==2.5.2
bottle==0.12.7
